requirements.txt
llama-index==0.10.36
llama-index-cli==0.1.12
llama-index-core==0.10.36
llama-index-legacy==0.9.48
llamaindex-py-client==0.1.19

#agents
llama-index-agent-openai==0.2.4

#llms
llama-index-llms-gemini==0.1.7
llama-index-multi-modal-llms-gemini==0.1.6
llama-index-llms-llama-cpp==0.1.3
#llama-index-llms-ollama
llama-index-llms-mistralai==0.1.12
llama-index-llms-openai==0.1.19
llama-index-llms-groq==0.1.3
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3

#embeddings
llama-index-embeddings-gemini==0.1.6
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.9

#indices
llama-index-indices-managed-llama-cloud==0.1.6

#readers
llama-index-readers-file==0.1.22
llama-index-readers-google==0.2.5
llama-index-readers-llama-parse==0.1.4
llama-index-readers-notion==0.1.6
llama-index-readers-web==0.1.13

#vector stores
llama-index-vector-stores-qdrant==0.2.8

#tools
llama-index-tools-tavily-research==0.1.3

#parsers
llama-parse==0.4.2

#other requirements
accelerate
bitsandbytes
docx2txt
EbookLib
fastparquet 
#flash-attn # LLaVA-Next; CUDA 11.8 only
html2text
jsonpath-ng>=1.6.1
llvmlite>=0.42.0
matplotlib
mistralai>=0.1.8
nbconvert
numpy 
openai>=1.21.2
openpyxl
pandas 
pillow
python-pptx
pyarrow
requests
sentencepiece
# torch # Omitted because this will re-install non-CUDA PyTorch
trafilatura>=1.9.0
transformers>=4.40.2
google-generativeai>=0.3.0
qdrant_client

#autogen
pyautogen>=0.2.27
tavily_python>=0.3.3